explicitly track Reset flag
this makes it easier to handle cases like \e[41m\e[0m\e[31m when
re-rendering the vterm. previously you would just see a format with only
a background, followed by a format with only a foreground, and you'd
have to deduce that a reset is needed. now you can just check for Reset
and print a reset first.

alternatively you could just _always_ reset when the format changes, but
that feels slightly more wasteful.