switch to golden style testing
lots of advantages to this. test failures are easy to understand, each
test covers a _ton_ of behavior, and it's trivial to go from repro =>
test.

only challenge is storing all the test fixtures. I started with
testing each frame (esc sequence), but for writing-readme that amounted
to 65k+ frames, which was ~500MB of space lol. so now it just tests each
Nth frame, but it's at least easy to override it in the tests to narrow
down the failure.