add AppendOnly mode
This is to support the "I just want to render log output" use case.

Without this flag it's somewhat easy for concurrent threads to step on
each other's toes and print something like "\e[95r" if one thread prints
"\e[95" and the other thread prints "random string".

With this mode we can avoid such an unfortunate series of events.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>